requirements.txt
streamlit
pandas
scikit-learn
pyyaml
bcrypt